+++ NEWS.md
@@ -3,7 +3,7 @@
 * Modified backend code
   * New `a_bool` function returns thresholded boolean matrix
 * Modified `[` method
-  * New `drop` argument toggles whether to use `simplify`
+  * New `tiny` argument toggles whether to use `simplify`
   * New `bool` argument toggles whether to use `a_bool`
 
 ## propr 2.0.4